Added count columns to posts table and structured some of the post action display logic to optimize view loading

+++ app/helpers/post_actions_helper.rb
@@ -1,20 +1,28 @@
 module PostActionsHelper
 
   def like_link(post)
-    like = post.likes.find_by_user_id(current_user)
-	  if like.nil?
-	    link_to("Like", likes_path(:post_id => post), :method => :post, :remote => true)
-	  else
-	    link_to("Unlike", like, :method => :delete, :remote => true)
-	 end
+    if post.likes_count?
+      like = post.likes.find_by_user_id(current_user)
+	    if like.nil?
+	      link_to("Like", likes_path(:post_id => post), :method => :post, :remote => true)
+	    else
+	      link_to("Unlike", like, :method => :delete, :remote => true)
+	   end
+	else
+	  link_to("Like", likes_path(:post_id => post), :method => :post, :remote => true)
+	end
   end
   
   def favorite_link(user, post)
-    favorite = post.favorites.find_by_user_id(user.id)
-	if favorite.nil?
-	  link_to((image_tag 'unfavorite.png', :height => "2.5%", :width => "2.5%"), favorites_path(:post_id => post), :method => :post, :remote => true)
+    if post.favorites_count?
+      favorite = post.favorites.find_by_user_id(user.id)
+	  if favorite.nil?
+	    link_to((image_tag 'unfavorite.png', :height => "2.5%", :width => "2.5%"), favorites_path(:post_id => post), :method > :post, :remote => true)
+	  else
+	    link_to((image_tag 'favorite.png', :height => "2.5%", :width => "2.5%"), favorite, :method => :delete, :remote => true)
+	  end
 	else
-	  link_to((image_tag 'favorite.png', :height => "2.5%", :width => "2.5%"), favorite, :method => :delete, :remote => true)
+	  link_to((image_tag 'unfavorite.png', :height => "2.5%", :width => "2.5%"), favorites_path(:post_id => post), :method > :post, :remote => true)
 	end
   end
